# speedup `batch` of arrays