--- a/src/utils.jl
+++ b/src/utils.jl
@@ -362,7 +362,7 @@ julia> unbatch([1 3 5 7;
  [5, 6]
  [7, 8]
 """
-unbatch(x::AbstractArray) = unstack(x, dims=ndims(x))
+unbatch(x::AbstractArray) = [getobs(x, i) for i in 1:numobs(x)]
 unbatch(x::AbstractVector) = x
 
 """