Previously `batch(xs)` was falling back to `stack(xs, dims=N+1)` that in turns calls `cat(xs...; dims=N+1)`. But for `stack` we can use `reduce(vcat, xs)` instead. 
```julia
using MLUtils, BenchmarkTools

function f(n)
    x = rand(Float32, 28, 28)
    return batch([x for i=1:n])
end

julia> @btime f(10);
  3.094 μs (28 allocations: 35.12 KiB)           # PR
  21.396 μs (139 allocations: 39.58 KiB)      # master

julia> @btime f(100);
  26.058 μs (208 allocations: 320.02 KiB)      # PR
  387.576 μs (1129 allocations: 433.50 KiB)   # master

julia> @btime f(1000);
  299.809 μs (2008 allocations: 3.09 MiB)    # PR
  27.570 ms (12023 allocations: 17.03 MiB)  # master
```